# api/openai/openapi/openai-minimal.yaml
openapi: 3.0.3
info:
  title: OpenAI-Compatible API
  description: Minimal OpenAI-compatible API for OmniAgent
  version: 1.0.0
servers:
  - url: /v1
paths:
  /chat/completions:
    post:
      operationId: createChatCompletion
      summary: Creates a model response for the given chat conversation
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatCompletionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateChatCompletionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /models/{model}:
    get:
      operationId: retrieveModel
      summary: Retrieves a model instance
      tags:
        - Models
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
          description: The ID of the model to retrieve
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    CreateChatCompletionRequest:
      type: object
      required:
        - model
        - messages
      properties:
        model:
          type: string
          description: ID of the model to use
          example: "omniagent"
        messages:
          type: array
          description: A list of messages comprising the conversation
          items:
            $ref: '#/components/schemas/ChatCompletionMessageParam'
          minItems: 1
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 1
          nullable: true
          description: Sampling temperature between 0 and 2
        top_p:
          type: number
          minimum: 0
          maximum: 1
          default: 1
          nullable: true
          description: Nucleus sampling probability
        n:
          type: integer
          minimum: 1
          maximum: 128
          default: 1
          nullable: true
          description: Number of chat completion choices to generate
        stream:
          type: boolean
          default: false
          nullable: true
          description: If true, partial message deltas will be sent as SSE
        stop:
          description: Stop sequences
          nullable: true
          oneOf:
            - type: string
            - type: array
              items:
                type: string
              maxItems: 4
        max_tokens:
          type: integer
          nullable: true
          description: Maximum number of tokens to generate
        max_completion_tokens:
          type: integer
          nullable: true
          description: Maximum number of tokens to generate (alternative to max_tokens)
        presence_penalty:
          type: number
          minimum: -2
          maximum: 2
          default: 0
          nullable: true
          description: Penalty for token presence
        frequency_penalty:
          type: number
          minimum: -2
          maximum: 2
          default: 0
          nullable: true
          description: Penalty for token frequency
        logit_bias:
          type: object
          additionalProperties:
            type: integer
          nullable: true
          description: Modify likelihood of specified tokens
        user:
          type: string
          description: Unique identifier for the end-user
        tools:
          type: array
          description: List of tools the model may call
          items:
            $ref: '#/components/schemas/ChatCompletionTool'
        tool_choice:
          description: Controls which tool is called by the model
          oneOf:
            - type: string
              enum: ["none", "auto", "required"]
            - $ref: '#/components/schemas/ChatCompletionNamedToolChoice'
        response_format:
          $ref: '#/components/schemas/ResponseFormat'
        seed:
          type: integer
          nullable: true
          description: Seed for deterministic sampling
        stream_options:
          $ref: '#/components/schemas/ChatCompletionStreamOptions'
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
          description: Custom metadata
    ChatCompletionMessageParam:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum: ["system", "user", "assistant", "tool"]
          description: The role of the message author
        content:
          description: The contents of the message
          oneOf:
            - type: string
            - type: array
              items:
                $ref: '#/components/schemas/ChatCompletionContentPart'
          nullable: true
        name:
          type: string
          description: Optional name for the participant
        tool_calls:
          type: array
          description: Tool calls generated by the model
          items:
            $ref: '#/components/schemas/ChatCompletionMessageToolCall'
        tool_call_id:
          type: string
          description: Tool call that this message is responding to
    ChatCompletionContentPart:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: ["text", "image_url"]
        text:
          type: string
        image_url:
          type: object
          properties:
            url:
              type: string
            detail:
              type: string
              enum: ["auto", "low", "high"]
              default: "auto"
    ChatCompletionMessageToolCall:
      type: object
      required:
        - id
        - type
        - function
      properties:
        id:
          type: string
          description: The ID of the tool call
        type:
          type: string
          enum: ["function"]
        function:
          type: object
          required:
            - name
            - arguments
          properties:
            name:
              type: string
              description: The name of the function to call
            arguments:
              type: string
              description: JSON arguments to pass to the function
    ChatCompletionTool:
      type: object
      required:
        - type
        - function
      properties:
        type:
          type: string
          enum: ["function"]
        function:
          $ref: '#/components/schemas/FunctionObject'
    FunctionObject:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: The name of the function
        description:
          type: string
          description: Description of what the function does
        parameters:
          type: object
          description: JSON Schema for the function parameters
        strict:
          type: boolean
          default: false
          nullable: true
          description: Whether to enable strict schema adherence
    ChatCompletionNamedToolChoice:
      type: object
      required:
        - type
        - function
      properties:
        type:
          type: string
          enum: ["function"]
        function:
          type: object
          required:
            - name
          properties:
            name:
              type: string
    ResponseFormat:
      type: object
      properties:
        type:
          type: string
          enum: ["text", "json_object", "json_schema"]
          default: "text"
        json_schema:
          type: object
          properties:
            name:
              type: string
            description:
              type: string
            schema:
              type: object
            strict:
              type: boolean
              default: false
    ChatCompletionStreamOptions:
      type: object
      nullable: true
      properties:
        include_usage:
          type: boolean
          description: If true, include usage information in the final chunk
    CreateChatCompletionResponse:
      type: object
      required:
        - id
        - object
        - created
        - model
        - choices
      properties:
        id:
          type: string
          description: Unique identifier for the completion
        object:
          type: string
          enum: ["chat.completion"]
        created:
          type: integer
          description: Unix timestamp of when the completion was created
        model:
          type: string
          description: The model used for completion
        choices:
          type: array
          items:
            $ref: '#/components/schemas/ChatCompletionChoice'
        usage:
          $ref: '#/components/schemas/CompletionUsage'
        system_fingerprint:
          type: string
          description: Backend configuration fingerprint
        service_tier:
          type: string
          nullable: true
          description: The service tier used
    ChatCompletionChoice:
      type: object
      required:
        - index
        - message
        - finish_reason
      properties:
        index:
          type: integer
          description: Index of this choice
        message:
          $ref: '#/components/schemas/ChatCompletionMessage'
        finish_reason:
          type: string
          enum: ["stop", "length", "tool_calls", "content_filter", "function_call"]
          nullable: true
          description: The reason the model stopped generating
        logprobs:
          $ref: '#/components/schemas/ChatCompletionLogprobs'
    ChatCompletionMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: ["assistant"]
        content:
          type: string
          nullable: true
          description: The contents of the message
        tool_calls:
          type: array
          items:
            $ref: '#/components/schemas/ChatCompletionMessageToolCall'
        function_call:
          type: object
          deprecated: true
          properties:
            name:
              type: string
            arguments:
              type: string
        refusal:
          type: string
          nullable: true
    ChatCompletionLogprobs:
      type: object
      nullable: true
      properties:
        content:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ChatCompletionTokenLogprob'
        refusal:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ChatCompletionTokenLogprob'
    ChatCompletionTokenLogprob:
      type: object
      required:
        - token
        - logprob
        - bytes
        - top_logprobs
      properties:
        token:
          type: string
        logprob:
          type: number
        bytes:
          type: array
          nullable: true
          items:
            type: integer
        top_logprobs:
          type: array
          items:
            type: object
            required:
              - token
              - logprob
              - bytes
            properties:
              token:
                type: string
              logprob:
                type: number
              bytes:
                type: array
                nullable: true
                items:
                  type: integer
    CompletionUsage:
      type: object
      required:
        - prompt_tokens
        - completion_tokens
        - total_tokens
      properties:
        prompt_tokens:
          type: integer
          description: Number of tokens in the prompt
        completion_tokens:
          type: integer
          description: Number of tokens in the completion
        total_tokens:
          type: integer
          description: Total number of tokens used
        prompt_tokens_details:
          type: object
          properties:
            cached_tokens:
              type: integer
            audio_tokens:
              type: integer
        completion_tokens_details:
          type: object
          properties:
            reasoning_tokens:
              type: integer
            audio_tokens:
              type: integer
            accepted_prediction_tokens:
              type: integer
            rejected_prediction_tokens:
              type: integer
...
    Model:
      type: object
      required:
        - id
        - object
        - created
        - owned_by
      properties:
        id:
          type: string
          description: The model identifier
        object:
          type: string
          enum: ["model"]
        created:
          type: integer
          description: Unix timestamp when the model was created
        owned_by:
          type: string
          description: Organization that owns the model
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
            - type
          properties:
            message:
              type: string
              description: Human-readable error message
            type:
              type: string
              description: Error type
            param:
              type: string
              nullable: true
              description: The parameter that caused the error
            code:
              type: string
              nullable: true
              description: Error code
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []